refactor(api): extract shared property create schema and helpers
Lifts createPropertyBodySchema, areSelectOptionsValid, createDefaultTool,
and the per-content-type content/tool builder out of create.ts into
create-schema.ts so a batch endpoint can reuse the same validation and
parts builder.
